[fix #6739] Do not send duplicate querystring on HTTP invocation
